feat(#869): remove redundant classes
